Move existing CTN composition into the shared Rust fit and prediction service
