+++ crates/gam-models/src/fit_orchestration/fit.rs
@@ -3734,318 +3734,6 @@ pub(crate) fn fit_transformation_normal_model(
     )
 }
 
-// ---------------------------------------------------------------------------
-// Cross-fitted score calibration (Neyman-orthogonal marginal slope, #461)
-//
-// A marginal-slope Stage-2 model consumes the CTN Stage-1 latent score `z` as a
-// generated regressor. `z` depends on θ̂₁, so the β estimating equation leaks
-// Stage-1 sampling error (design `marginal_slope_orthogonal_design.md` §1-§4).
-// The two DML ingredients are (i) the realized leakage directions
-// `J = ∂z/∂θ₁` (an n × p₁ influence Jacobian, computed by the core
-// `marginal_slope_orthogonal::score_influence_jacobian`) and (ii) cross-fitting:
-// `z` and `J` are evaluated out-of-fold so own-row overfitting of θ̂₁ does not
-// bias the absorbed projection. This module owns ingredient (ii) and the
-// auto-enable detection: it partitions the rows into K folds, refits the CTN on
-// each complement with a basis frozen on the full data, and concatenates the
-// per-fold held-out `z` and `J` back into full-n order.
-
-/// Number of cross-fit folds for a problem of `n` rows.
-///
-/// Cross-fitting refits the CTN once per fold, so the cost is `K` full Stage-1
-/// fits. The standard DML default is `K = 5` for moderate `n`. At large scale
-/// each CTN refit is expensive while the out-of-fold bias from a single split is
-/// already negligible (θ̂₁ is precisely estimated on a complement of ≈ `n·(K−1)/K`
-/// rows), so `K` is reduced toward 2 as `n` grows — keeping the refit budget
-/// bounded without sacrificing OOF de-biasing. Small `n` keeps more folds (larger
-/// per-fold training complements ⇒ less OOF estimation noise), dropping below 5
-/// only when there are too few rows to populate 5 folds with a usable held-out
-/// block.
-///
-/// The schedule (no flag, no env — derived purely from `n`):
-///   - `n < 250`               : `K = min(n, 3)` (tiny data; keep ≥ 2 folds)
-///   - `250 ≤ n < 200_000`      : `K = 5` (DML moderate-n default)
-///   - `200_000 ≤ n < 2_000_000` : `K = 3` (large-scale: bound refit cost, ≈ ⅔ train)
-///   - `n ≥ 2_000_000`          : `K = 2` (mega-large-scale: ½ train still ample)
-fn crossfit_fold_count(n: usize) -> usize {
-    if n < 250 {
-        n.min(3).max(2)
-    } else if n < 200_000 {
-        5
-    } else if n < 2_000_000 {
-        3
-    } else {
-        2
-    }
-}
-
-/// Partition `n` rows into `k` folds of balanced, contiguous blocks.
-///
-/// Entry `f` of the returned vector holds the ascending row indices held out in
-/// fold `f`; the union over folds is exactly `0..n`. Sizes are `n / k` or
-/// `n / k + 1`, so every fold's complement size differs by at most one row,
-/// which keeps the response-basis sample cap — and therefore `p₁` — uniform
-/// across folds (design §3). Contiguous blocks let the persistent warm-start
-/// prefix cache seed each fold's CTN refit from a structurally identical prior
-/// fold.
-fn crossfit_partition(n: usize, k: usize) -> Vec<Vec<usize>> {
-    let mut folds: Vec<Vec<usize>> = Vec::with_capacity(k);
-    let base = n / k;
-    let remainder = n % k;
-    let mut start = 0usize;
-    for f in 0..k {
-        // The first `remainder` folds carry one extra row so the union is exactly n.
-        let len = base + usize::from(f < remainder);
-        let end = start + len;
-        folds.push((start..end).collect());
-        start = end;
-    }
-    folds
-}
-
-/// Gather `source[idx]` for each `idx` into a fresh contiguous `Array1`.
-fn crossfit_select_rows_1d(source: &Array1<f64>, indices: &[usize]) -> Array1<f64> {
-    Array1::from_iter(indices.iter().map(|&i| source[i]))
-}
-
-/// Cross-fitted out-of-fold `z` and score-influence Jacobian `J` for a CTN →
-/// marginal-slope chain (design §4-§5).
-///
-/// Returns `None` when no CTN Stage-1 recipe is present (`recipe` is `None`):
-/// the caller then leaves the Stage-2 spec's `score_influence_jacobian` field
-/// `None` and Stage-2 uses the supplied raw `z` with the free-warp fallback.
-///
-/// When a recipe is present, the covariate basis is built from the recipe's
-/// formula RHS and FROZEN once on the full data; every fold then refits the CTN
-/// against that frozen spec, and the response-basis knot count is pre-resolved at
-/// the *smallest* fold complement size, so every fold refits at an identical
-/// `(p_resp, p_cov)` and therefore an identical `p₁ = p_resp · p_cov` column
-/// layout — the per-fold `J` blocks concatenate into a coherent `n × p₁` matrix
-/// (design §3). For each fold `f` the CTN is refit on the complement rows, then
-/// `marginal_slope_orthogonal::score_influence_jacobian` evaluates the held-out
-/// `z` and `J` on fold `f`'s rows; results scatter back into full-n order.
-pub(crate) fn crossfit_score_calibration(
-    data: &Dataset,
-    col_map: &HashMap<String, usize>,
-    recipe: Option<&CtnStage1Recipe>,
-    policy: &gam_runtime::resource::ResourcePolicy,
-) -> Result<Option<CrossFitScoreCalibration>, String> {
-    let Some(recipe) = recipe else {
-        return Ok(None);
-    };
-
-    let n = data.values.nrows();
-    if n == 0 {
-        return Err("cross-fit score calibration requires a non-empty dataset".to_string());
-    }
-
-    // Stage-1 response / weights / offset, resolved against the full dataset.
-    let y_col = resolve_role_col(col_map, &recipe.response_column, "response")
-        .map_err(|e| e.to_string())?;
-    let response_full = data.values.column(y_col).to_owned();
-    let weights_full = resolve_weight_column(data, col_map, recipe.weight_column.as_deref())
-        .map_err(|e| e.to_string())?;
-    let offset_full = resolve_offset_column(data, col_map, recipe.offset_column.as_deref())
-        .map_err(|e| e.to_string())?;
-
-    // Build the CTN covariate basis from the recipe's formula RHS and FREEZE it
-    // ONCE on full data, so every fold refit reuses identical spatial centers /
-    // knots ⇒ identical p_cov across folds (design §3). The freeze is what makes
-    // the per-fold covariate designs column-aligned.
-    let parsed_cov = parse_formula(&format!(
-        "{} ~ {}",
-        recipe.response_column, recipe.covariate_formula_rhs
-    ))
-    .map_err(|e| e.to_string())?;
-    let mut frozen_notes = Vec::new();
-    let mut covariate_spec_raw = build_termspec_with_geometry_and_overrides(
-        &parsed_cov.terms,
-        data,
-        col_map,
-        &mut frozen_notes,
-        false,
-        policy,
-        None,
-        None,
-    )
-    .map_err(|e| e.to_string())?;
-    // #2750/#2754: resolve auto measure-jet ranges HERE, once, on the full data
-    // — before the freeze below and therefore before any fold sees the spec.
-    //
-    // This is the one place in the cross-fit path where the screen can run at
-    // all. Every fold refit is handed `frozen_cov_spec`, and a frozen term is a
-    // replay with nothing left to seed, so the screen inside
-    // `fit_transformation_normal` is correctly a no-op there. Screening on the
-    // full response before the freeze keeps the property the freeze exists for
-    // — identical centers, identical `p_cov`, column-aligned designs across
-    // folds — because every fold then replays the SAME screened range, rather
-    // than each choosing its own from its own complement.
-    let seeded = crate::fit_orchestration::drivers::seed_measure_jet_auto_ranges(
-        data.values.view(),
-        response_full.view(),
-        weights_full.view(),
-        &mut covariate_spec_raw,
-    );
-    if seeded > 0 {
-        log::info!(
-            "[#2750] screened the representer range of {seeded} auto measure-jet term(s) against \
-             the full response before freezing the cross-fit Stage-1 covariate basis"
-        );
-    }
-    let covariate_spec_raw = covariate_spec_raw;
-    let full_cov_design = build_term_collection_design(data.values.view(), &covariate_spec_raw)
-        .map_err(|e| e.to_string())?;
-    let frozen_cov_spec = crate::fit_orchestration::drivers::freeze_term_collection_from_design(
-        &covariate_spec_raw,
-        &full_cov_design,
-    )
-    .map_err(|e| e.to_string())?;
-    let p_cov = full_cov_design.design.ncols();
-
-    let k = crossfit_fold_count(n);
-    let folds = crossfit_partition(n, k);
-
-    // Pre-resolve the response-basis internal-knot count at the *smallest* fold
-    // complement size. The CTN sample cap on this count is monotone in the
-    // complement size, so pinning the per-fold config to the value resolved at
-    // the smallest complement makes every fold resolve to the same count —
-    // hence a fold-invariant p_resp and an aligned p₁ across folds (design §3).
-    let min_complement = folds.iter().map(|held| n - held.len()).min().unwrap_or(n);
-    let mut fold_config = recipe.config.clone();
-    fold_config.response_num_internal_knots =
-        crate::transformation_normal::effective_response_num_internal_knots(
-            &recipe.config,
-            min_complement,
-            p_cov,
-            response_full.view(),
-        );
-    // Pin the resolved knot count: each fold's CTN refit must use exactly this
-    // value, not re-derive it from its own response subsample. Without this the
-    // data-driven complexity cap rounds to different counts per fold, so p_resp
-    // (and p₁ = p_resp · p_cov) drifts across folds and the OOF Jacobian
-    // assembly fails the fold-alignment check below ("cross-fit fold p₁ mismatch").
-    fold_config.response_num_internal_knots_pinned = true;
-    // Resolve the response knot vector ONCE on the FULL response, exactly as the
-    // covariate side is frozen once on full data above, and pin it into every
-    // fold (gam#2680).
-    //
-    // The knot vector fixes the certified response support `[y_lo, y_hi]`, and a
-    // CTN's score is the PIT of `h(y|x)` against a standard normal TRUNCATED to
-    // that support. Left fold-local it breaks the out-of-fold score twice over:
-    //
-    //   1. whichever fold holds out the global response minimum or maximum
-    //      evaluates it against a support built from a complement that does not
-    //      contain it, so the held-out row is outside its own fold's certified
-    //      domain and `transformation_normal_pit_score` refuses — taking the
-    //      entire OOF assembly with it. With `K` folds and a guard of 0.1 % of
-    //      the fold span that is close to guaranteed;
-    //   2. even with no refusal, `z_oof` would be stitched together from `K`
-    //      PITs taken against `K` different truncations. The Stage-2 consumers
-    //      require `z ~ N(0, 1)` and the moment gate in `bms::gradient_paths`
-    //      refuses a score that is not — but a mixture of differently-truncated
-    //      PITs is not one latent scale at all, so the gate would be measuring a
-    //      quantity that has no single definition.
-    //
-    // The full response is the right seed: every fold's held-out rows are drawn
-    // from it, so the support contains all of them by construction, and the
-    // knots are identical across folds, so `p_resp` is fold-invariant for a
-    // second, structural reason on top of the count pin.
-    fold_config.response_knots_pinned = Some(crate::transformation_normal::ctn_response_knots(
-        response_full.view(),
-        fold_config.response_degree,
-        fold_config.response_num_internal_knots,
-    )?);
-
-    let mut z_oof = Array1::<f64>::zeros(n);
-    let mut jac_oof: Option<Array2<f64>> = None;
-
-    for held in &folds {
-        if held.is_empty() {
-            continue;
-        }
-        let held_set: std::collections::HashSet<usize> = held.iter().copied().collect();
-        let complement: Vec<usize> = (0..n).filter(|i| !held_set.contains(i)).collect();
-        if complement.is_empty() {
-            return Err(
-                "cross-fit fold left an empty training complement; too few rows for K folds"
-                    .to_string(),
-            );
-        }
-
-        // Refit the CTN on the complement (training) rows. The covariate design
-        // it uses comes from the frozen spec, so its column geometry matches the
-        // held-out Jacobian evaluation below and every other fold.
-        let train_cov = data.values.select(Axis(0), &complement);
-        let train_resp = crossfit_select_rows_1d(&response_full, &complement);
-        let train_weights = crossfit_select_rows_1d(&weights_full, &complement);
-        let train_offset = crossfit_select_rows_1d(&offset_full, &complement);
-
-        let fold_fit = fit_transformation_normal(
-            &train_resp,
-            &train_weights,
-            &train_offset,
-            train_cov.view(),
-            &frozen_cov_spec,
-            &fold_config,
-            &BlockwiseFitOptions::default(),
-            &SpatialLengthScaleOptimizationOptions::default(),
-            None,
-        )?;
-
-        // Evaluate the OOF score z and the OOF influence Jacobian J on fold f's
-        // held-out rows from this fold's fitted CTN. The held-out offset rows
-        // enter the PIT operating point identically to the Stage-1 row build, so
-        // the emitted z matches the fitted model when the recipe carries an
-        // offset column (zeros otherwise ⇒ no-op).
-        let held_cov = data.values.select(Axis(0), held);
-        let held_resp = crossfit_select_rows_1d(&response_full, held);
-        let held_offset = crossfit_select_rows_1d(&offset_full, held);
-
-        let jac = crate::marginal_slope_orthogonal::score_influence_jacobian(
-            &fold_fit,
-            &held_resp,
-            held_cov.view(),
-            &held_offset,
-        )?;
-
-        if jac.columns.nrows() != held.len() {
-            return Err(format!(
-                "cross-fit fold Jacobian row count {} != held-out fold size {}",
-                jac.columns.nrows(),
-                held.len()
-            ));
-        }
-        if jac.z.len() != held.len() {
-            return Err(format!(
-                "cross-fit fold OOF z length {} != held-out fold size {}",
-                jac.z.len(),
-                held.len()
-            ));
-        }
-
-        let p1 = jac.columns.ncols();
-        let jac_full = jac_oof.get_or_insert_with(|| Array2::<f64>::zeros((n, p1)));
-        if jac_full.ncols() != p1 {
-            return Err(format!(
-                "cross-fit fold p₁ mismatch: this fold has {p1} columns but a prior fold had {}; \
-                 the frozen response/covariate basis failed to align across folds",
-                jac_full.ncols()
-            ));
-        }
-
-        for (local, &global) in held.iter().enumerate() {
-            z_oof[global] = jac.z[local];
-            for c in 0..p1 {
-                jac_full[[global, c]] = jac.columns[[local, c]];
-            }
-        }
-    }
-
-    let jac_oof = jac_oof.ok_or_else(|| {
-        "cross-fit produced no folds with held-out rows; cannot assemble OOF Jacobian".to_string()
-    })?;
-
-    Ok(Some(CrossFitScoreCalibration { z_oof, jac_oof }))
-}
 
 #[cfg(test)]
 mod survival_edf_tests {
