+++ crates/gam-models/src/fit_orchestration/request.rs
@@ -343,6 +343,8 @@ pub struct SurvivalTransformationFitResult {
 }
 
 pub enum FitResult {
+    /// Complete CTN/outcome predictor with frozen native transformation state.
+    Ctn(Box<crate::inference::model::FittedModelPayload>),
     Standard(StandardFitResult),
     GaussianLocationScale(GaussianLocationScaleFitResult),
     BinomialLocationScale(BinomialLocationScaleFitResult),
@@ -391,49 +393,23 @@ pub struct DispersionLocationScaleFitResult {
     pub kind: DispersionFamilyKind,
 }
 
-/// Out-of-fold Stage-1 latent score and its score-influence Jacobian for a
-/// CTN → marginal-slope chain. `z_oof` (length n) replaces the in-sample `z`
-/// the Stage-2 model consumes; `jac_oof` (n × p₁) is fed to the Stage-2 spec's
-/// `score_influence_jacobian` so the joint solve absorbs the realized leakage
-/// directions `Z_infl = diag(s_f·β̂₀)·J`.
-pub struct CrossFitScoreCalibration {
-    pub z_oof: Array1<f64>,
-    pub jac_oof: Array2<f64>,
-}
-
-/// Internal recipe describing the CTN Stage-1 fit that produced a Stage-2 `z`
-/// column. This is in-process plumbing — never a CLI flag, env var, or feature
-/// gate. The orchestration layer populates [`FitConfig::ctn_stage1`] when (and
-/// only when) the marginal-slope `z` was generated by a transformation-normal
-/// Stage-1 fit; its presence is the sole auto-enable signal for cross-fitted
-/// orthogonalization (design §5). When absent, Stage-2 falls back to the free
-/// 1-D `score_warp` spline (which spans only the x-free leakage column).
+/// Native cross-fitted conditional score recipe. Each fold learns its own
+/// geometry; a full-training CTN is saved for deployment. No influence absorber.
 #[derive(Clone, Debug)]
 pub struct CtnStage1Recipe {
+    /// Explicit nuisance-fold column, optionally checked against groups.
+    pub fold_column: Option<String>,
+    /// Independent groups used for reproducible whole-group folds.
+    pub group_column: Option<String>,
+    /// Number of generated group folds.
+    pub folds: usize,
+    /// Reproducible group-fold seed.
+    pub seed: u64,
     /// Stage-1 response column name (the `y` the CTN transforms).
     pub response_column: String,
-    /// Stage-1 covariate-side formula right-hand side (e.g. `"s(pc1) + s(pc2)"`),
-    /// with no `~` and no response symbol. `crossfit_score_calibration` parses
-    /// it and builds the CTN covariate basis exactly as
-    /// `materialize_transformation_normal` does, then FREEZES that basis once on
-    /// the full data and reuses the frozen spec for every fold's refit — so the
-    /// rebuilt covariate design has an identical column geometry across folds,
-    /// keeping `J`'s `p₁ = p_resp · p_cov` columns aligned (design §3).
-    ///
-    /// The recipe carries the formula RHS (a primitive string) rather than a
-    /// resolved [`TermCollectionSpec`] because this struct is populated both via
-    /// [`CtnStage1Recipe::new`] (set on [`FitConfig::ctn_stage1`], then
-    /// [`fit_from_formula`]) and by the gamfit FFI marshaller
-    /// (`gamfit/_calibrated_slope.py`), which can only serialize primitives over
-    /// the JSON boundary — a `TermCollectionSpec` is not serializable. Freezing on
-    /// the full Stage-2 data is equivalent to
-    /// freezing on the Stage-1 data whenever the two stages share a frame (the
-    /// calibrated-chain contract), so the column geometry still matches Stage-1.
+    /// Covariate formula RHS. Each nuisance fold learns its own geometry.
     pub covariate_formula_rhs: String,
-    /// Stage-1 CTN config (response basis degree / knot count / penalties).
-    /// Its `response_num_internal_knots` is the FIXED response-basis size; the
-    /// cross-fit pins it across folds so `p_resp` (and hence `p₁`) is
-    /// fold-invariant (design §3).
+    /// Existing CTN response basis and function-penalty configuration.
     pub config: TransformationNormalConfig,
     /// Optional Stage-1 weight column name.
     pub weight_column: Option<String>,
@@ -443,11 +419,9 @@ pub struct CtnStage1Recipe {
 
 impl CtnStage1Recipe {
     /// Build a Stage-1 CTN recipe from the Stage-1 description. This is the public
-    /// way to populate [`FitConfig::ctn_stage1`] — set it on a marginal-slope
-    /// config and run [`fit_from_formula`] (the entry IS `fit_from_formula` with
-    /// `ctn_stage1` set; there is no separate combined entry function). The
-    /// materializer then cross-fits the CTN and installs the leakage-projection
-    /// block; supplying the recipe *is* the request for orthogonalization.
+    /// way to populate [`FitConfig::ctn_stage1`]. Supply fold/group columns and
+    /// call [`fit_from_formula`] for the complete native predictor. Cross-fitting
+    /// does not install an influence absorber or establish orthogonality.
     ///
     /// `response` is the Stage-1 CTN response column; `covariates` is the
     /// covariate-side formula right-hand side (e.g. `"s(pc1) + s(pc2)"` — no `~`,
@@ -478,6 +452,10 @@ impl CtnStage1Recipe {
             );
         }
         Ok(Self {
+            fold_column: None,
+            group_column: None,
+            folds: 5,
+            seed: 0,
             response_column,
             covariate_formula_rhs,
             config,
@@ -574,7 +552,8 @@ pub struct FitConfig {
     /// Column name for the z (exposure/dose) variable in marginal-slope models.
     pub z_column: Option<String>,
     /// Consume an externally fitted latent score without fitting another
-    /// conditional or rank-based transform. Distribution checks remain on.
+    /// conditional or rank-based transform. This assumes a standard-normal
+    /// latent law; freezing an input does not certify its conditional law.
     pub frozen_score: bool,
     /// Standalone CTN response-basis options, also used by predictive cross-fitting.
     pub transformation_normal_config: Option<TransformationNormalConfig>,
@@ -593,18 +572,11 @@ pub struct FitConfig {
     /// ignored for every other family. The asymmetry may also be written inline
     /// as `family = "expectile(0.9)"`, which fills this field at resolve time.
     pub expectile_tau: Option<f64>,
-    /// Internal CTN Stage-1 provenance for the marginal-slope `z` column.
-    ///
-    /// When the marginal-slope `z` was generated by a transformation-normal
-    /// Stage-1 fit, the orchestration layer fills this with the Stage-1 recipe.
-    /// Its presence is the sole auto-enable signal for cross-fitted, Neyman-
-    /// orthogonal score calibration (#461): the materializer cross-fits the CTN
-    /// to produce out-of-fold `z` and the score-influence Jacobian `J`, replaces
-    /// the raw `z` with `z_oof`, and absorbs `J` as a leakage-projection block in
-    /// Stage-2. This is in-process plumbing only — there is no CLI flag, env var,
-    /// or feature gate. `None` ⇒ raw `z` with the free-warp `score_warp`
-    /// fallback. See [`CtnStage1Recipe`].
+    /// Cross-fitted predictive CTN, saved with an ordinary marginal-slope outcome.
     pub ctn_stage1: Option<CtnStage1Recipe>,
+    /// A previously fitted CTN, applied unchanged to training and prediction.
+    /// Mutually exclusive with cross-fitting and an external latent-score column.
+    pub frozen_ctn: Option<crate::inference::ctn::FrozenCtn>,
 
     // Fitting options
     pub scale_dimensions: bool,
@@ -774,6 +746,7 @@ impl Default for FitConfig {
             weight_column: None,
             expectile_tau: None,
             ctn_stage1: None,
+            frozen_ctn: None,
             scale_dimensions: false,
             spatial_optimization: SpatialLengthScaleOptimizationOptions::default(),
             transformation_normal: false,
