+++ crates/gam-models/src/inference/mod.rs
@@ -6,6 +6,8 @@ pub mod model;
 pub mod model_extension;
 pub mod model_payload_builders;
 pub mod predict_io;
+pub mod predict_input;
+pub mod ctn;
 
 #[cfg(test)]
 mod marginal_slope_predict_tests;
