+++ crates/gam-pyffi/src/model/model_ffi.rs
@@ -2008,11 +2008,38 @@ fn predict_table(
     })
 }
 
+#[pyfunction]
+fn ctn_required_fit_columns(formula: String, config_json: String) -> PyResult<Vec<String>> {
+    let config = parse_fit_config(Some(&config_json)).map_err(py_value_error)?;
+    gam::inference::ctn::required_fit_columns(&formula, &config)
+        .map(|columns| columns.into_iter().collect()).map_err(py_value_error)
+}
+
+#[pyfunction]
+fn required_model_columns(model_bytes: Vec<u8>, observed_score: bool) -> PyResult<Vec<String>> {
+    let mut model = load_model_impl(&model_bytes).map_err(|error| py_value_error(error.to_string()))?;
+    if observed_score {
+        if let Some(transform) = model.score_transform.as_ref() {
+            model = FittedModel::from_payload((**transform).clone());
+        }
+    }
+    let mut columns = model.prediction_required_columns().map_err(py_value_error)?;
+    if observed_score {
+        let response = response_column_name(&model.formula)
+            .ok_or_else(|| py_value_error("CTN requires a named observed response"))?;
+        columns.insert(response);
+    }
+    Ok(columns.into_iter().collect())
+}
+
 fn transformation_score_encoded_table_impl(
     model_bytes: &[u8],
     source: EncodedDataset,
 ) -> Result<Array1<f64>, String> {
-    let model = load_model_impl(model_bytes).map_err(|error| error.to_string())?;
+    let mut model = load_model_impl(model_bytes).map_err(|error| error.to_string())?;
+    if let Some(transform) = model.score_transform.as_ref() {
+        model = FittedModel::from_payload((**transform).clone());
+    }
     if model.predict_model_class() != PredictModelClass::TransformationNormal {
         return Err(format!(
             "transformation_score requires a transformation-normal model; got '{}'",
