+++ gamfit/__init__.py
@@ -37,7 +37,6 @@ from pathlib import Path
 from ._api import (
     SUPPORT_SAE_SCHEMA,
     CtnStage1,
-    CtnMarginalSlopeModel,
     SharedPrecisionGroup,
     bspline_basis,
     bspline_basis_derivative,
